% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alignAndCropImages.R
\name{alignAndCropImages}
\alias{alignAndCropImages}
\title{Crop and Align A Series of Images}
\usage{
alignAndCropImages(phase, dyes, alignmentSample = c(200, 200, 100),
  cropBoundaries = c(50, 50, 50, 50), rotation = 0)
}
\arguments{
\item{phase}{a list of phase images}

\item{dyes}{a list of lists dye images of equal length to phase}

\item{alignmentSample}{region of background image to align with (in order:
starting x position, starting y position, square width)}

\item{cropBoundaries}{how many pixels to crop from each side (in order:
bottom, left, top, right}

\item{rotation}{degrees to rotate images}
}
\value{
a \code{list} of two lists, \code{phase} and \code{dyes},
which will be the same lengths as the input.
}
\description{
Given a list of phase images and a list of zero or
more lists of dye images, aligns these images to the first image
of phase (assumed to be a background). Will also crop and rotate
the images if those are specified.
}

