% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis_doublingTime.R
\name{analysis_doublingTime}
\alias{analysis_doublingTime}
\title{Calculate Doubling Times}
\usage{
analysis_doublingTime(timeseries, minExpFitHour = 0, maxExpFitHour = 1e+09)
}
\arguments{
\item{timeseries}{timeseries dataframe}

\item{minExpFitHour}{hour of first datapoint to include in doubling time exponential fit}

\item{maxExpFitHour}{hour of last datapoint to include in doubling time exponential fit}
}
\value{
Vector of doubling times
}
\description{
Given a "timeseries" dataframe with hours in the
first column and colony sizes in all other columns, return a
vector of doubling times in hours.
}

