% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis_winnowColonies.R
\name{analysis_winnowColonies}
\alias{analysis_winnowColonies}
\title{Remove Likely Artifacts}
\usage{
analysis_winnowColonies(timeseries, minExpFitHour = 0,
  maxExpFitHour = 1e+09, minDoublingTime = NULL, maxDoublingTime = NULL,
  removeOutliers = TRUE, maxStartHour = NULL)
}
\arguments{
\item{timeseries}{timeseries dataframe}

\item{minExpFitHour}{hour of first datapoint to include in doubling time exponential fit}

\item{maxExpFitHour}{hour of last datapoint to include in doubling time exponential fit}

\item{minDoublingTime}{colonies with doubling times smaller than this many hours are removed}

\item{maxDoublingTime}{colonies with doubling times larger than this many hours are removed}

\item{removeOutliers}{flag indicating whether to remove colonies with doubling time outliers}

\item{maxStartHour}{colonies not identified by this hour are removed}
}
\value{
List of two timeseries dataframes: "retained" and "removed".
}
\description{
After image analysis, the TBCellGrowth package produces.csv files
which can be read in as a "timeseries" dataframe
with hours in the first column and colony sizes in all other columns.
Depending on run configuration, some portion of the colonies will be artifacts
-- bubbles, blotches, specks, etc.

This function helps to winnow down likely artifacts and returns a dataframe
that likely contains a higher percentage of actual colonies.

The main way this function removes artifacts is by removing "colonies" that
do not grow. Doubling times for rapidly growing colonies will be low numbers
while doubling times for no or little growth will be very large numbers.
Negative doubling times are found where identified "colonies" shrink over time.

This function identifies any colony with a negative doubling time as an artifact.
After removal of negative doubling times, the boxlplot() function is used to
identifiy outliers. These are subsequently removed.

An additional option is to remove all "colonies" that are not identified until
after a particular hour. (Sometimes, changes in the image quality cause new
"colonies" to be identified after a particular timestep.)

This funciton returns a list with two dataframes, one containing the retained
colonies and one containing the removed colonies.
}
\details{
The R boxplot() function is used to define outliers and has the following documentation
where IQR stands for Inter-Quartile Range:

The two 'hinges' are versions of the first and third quartile, i.e., close to
quantile(x, c(1,3)/4). The hinges equal the quartiles for odd n (where n <- length(x))
and differ for even n. Whereas the quartiles only equal observations for n %% 4 == 1 (n = 1 mod 4),
the hinges do so additionally for n %% 4 == 2 (n = 2 mod 4), and are in the middle
of two observations otherwise.

The notches (if requested) extend to +/-1.58 IQR/sqrt(n). This seems to be based on
the same calculations as the formula with 1.57 in Chambers et al (1983, p. 62), given
in McGill et al (1978, p. 16). They are based on asymptotic normality of the median
and roughly equal sample sizes for the two medians being compared, and are said to
be rather insensitive to the underlying distributions of the samples. The idea appears
to be to give roughly a 95% confidence interval for the difference in two medians.
}
\seealso{
\link{analysis_doublingTime}
}

