% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/applyCropRect.R
\name{applyCropRect}
\alias{applyCropRect}
\title{Crop an Image}
\usage{
applyCropRect(image, width, height)
}
\arguments{
\item{image}{an image matrix}

\item{width}{twice the individual width border}

\item{height}{twice the individual height border}
}
\value{
Cropped image.
}
\description{
Removes border pixels from an image.
}

