% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildDirectoryStructure.R
\name{buildDirectoryStructure}
\alias{buildDirectoryStructure}
\title{Builds A Directory of Image With Excel UI}
\usage{
buildDirectoryStructure(output, phase, labeled, dyeOverlap, filenames,
  outputDir = "output")
}
\arguments{
\item{output}{an output object with timeseries and centroids from
\link{generateBlobTimeseries}}

\item{phase}{a list of phase contrast image matrices}

\item{dyeOverlap}{a \code{dataframe} of dye overlap values from
\link{findDyeOverlap}}

\item{filenames}{a vector of names equal to the length of the
image lists. Usually timesteps.}

\item{outputDir}{the directory to build the file structure}

\item{phase.labeled}{a list of labeled phase images from
\link{flow_labelPhase}}

\item{dyes.labeled}{a list of labeled dye images from
\link{flow_labelDye}}
}
\value{
none
}
\description{
write a lot more here TODO
}

