% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createArtifactMask.R
\name{createArtifactMask}
\alias{createArtifactMask}
\title{Creates a mask of non biological artifacts, such as pillars and squares.}
\usage{
createArtifactMask(bg, maskDarkLines = FALSE)
}
\arguments{
\item{bg}{a preprocessed background image.}

\item{maskDarkLines}{whether or not dark lines should be masked out.}
}
\value{
A \code{matrix} of boolean values with TRUE values identifying
non biological artifacts.
}
\description{
Creates a mask of non biological artifacts, such as pillars and squares.
}

