% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createGif.R
\name{createGif}
\alias{createGif}
\title{Creates a gif animation.}
\usage{
createGif(dir, filename, ext = "jpg", framerate = 2, rescale = 80)
}
\arguments{
\item{dir}{the directory of images to read}

\item{filename}{name of output gif.}

\item{ext}{the image file extension to read.}

\item{framerate}{the number of frames per second in the output gif.}

\item{rescale}{dimensions of gif frames as a percent of original image size.}
}
\value{
none
}
\description{
Creates an animated .gif file from a list of images using ImageMagick.
}

