% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cropImageByID.R
\name{cropImageByID}
\alias{cropImageByID}
\title{Crops an Image Around a Given Blob ID}
\usage{
cropImageByID(id, output, bg, label)
}
\arguments{
\item{bg}{a background image}

\item{label}{a binary image label}

\item{color}{a color string, i.e. "green" or "phase"}
}
\value{
a list with elements bg, the cropped background, and label, the cropped
label
}
\description{
Given a blob ID string, crops an image around that id.
}

