% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_blur.R
\name{filter_blur}
\alias{filter_blur}
\title{Blur an Image}
\usage{
filter_blur(image, radius = 3)
}
\arguments{
\item{image}{an image matrix}

\item{radius}{number of pixels to blur. Must be an odd number.}
}
\value{
An image of the same dimensions.
}
\description{
Blurs an image using the EBImage makeBrush and filter2
functions.
}

