% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter_LoG.R
\name{filter_LoG}
\alias{filter_LoG}
\title{Sobel Edge Detection Filter}
\usage{
filter_LoG(image)
}
\arguments{
\item{image}{image matrix to detect edges on}
}
\value{
an image of the same dimensions.
}
\description{
This is an implementation of the 3x3 sobel filter which
is used by Photoshop and EBImage.
}

