% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_sobel.R
\name{filter_sobel}
\alias{filter_sobel}
\title{Sobel Edge Detection Filter}
\usage{
filter_sobel(image)
}
\arguments{
\item{image}{image matrix to detect edges on}
}
\value{
an image of the same dimensions.
}
\description{
This is an implementation of the 3x3 sobel filter which
is used by Photoshop and EBImage.
}

