% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findDarkLines.R
\name{findDarkLines}
\alias{findDarkLines}
\title{Find Horizontal Dark Lines}
\usage{
findDarkLines(image)
}
\arguments{
\item{image}{matrix of image data}
}
\value{
A \code{matrix} of start and end indices for dark lines.
}
\description{
In Fluidomics apparatus, the space between plate levels often
generates dark lines in phase contrast microscopy.
}

