% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findDyeOverlap.R
\name{findDyeOverlap}
\alias{findDyeOverlap}
\title{Find the Amount of Overlap Between Dye and Phase Blobs}
\usage{
findDyeOverlap(dye, phase.labeled, output)
}
\arguments{
\item{dye}{a list of labeled dye images.}

\item{output}{}

\item{phase}{a list of labeled phase images.}
}
\value{
A \code{matrix} of integer labeled blobs.
}
\description{
Searches an image for dark cell colonies and incrementally labels each blob.
}

