% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_createArtifactMask.R
\name{flow_createArtifactMask}
\alias{flow_createArtifactMask}
\title{Create a Mask of non-Biological Artifacts}
\usage{
flow_createArtifactMask(bg, maskDarkLines = FALSE)
}
\arguments{
\item{bg}{preprocessed background image}

\item{maskDarkLines}{logical specifying whether or not dark lines should be masked out}
}
\value{
A \code{matrix} of boolean values with TRUE values identifying non biological artifacts.
}
\description{
Fluidomics apparatus used in phase microscopy have a number of physical
features that must be masked out before image processing algorithms to identify bacteria
are applied. This function is tailored to remove *pillars*, *squares* and the dark lines
that appear between different levels in the apparatus.
}

