% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow_equalizePhase.R
\name{flow_equalizePhase}
\alias{flow_equalizePhase}
\title{Normalize Image Values}
\usage{
flow_equalizePhase(image, medianNew)
}
\arguments{
\item{image}{the image matrix to modify.}

\item{medianNew}{a new median value to set scale the image values against}
}
\value{
an image of the same dimensions.
}
\description{
Normalize a given image matrix based on median. The new median
 value should be between 0 and 1, and will typically be between 0.2 and 0.6.
}

