% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_findIgnore.R
\name{flow_findIgnore}
\alias{flow_findIgnore}
\title{Find Boundaries of Ignored Areas}
\usage{
flow_findIgnore(ignore, dims)
}
\arguments{
\item{ignore}{a vector of ignore strings (i.e. "topRight")}

\item{dims}{the dimensions of the image}
}
\value{
A \code{data frame} of rectangle boundaries corresponing to
ignore regions.
}
\description{
In some cases, Fluidomics image stitched together with
9 images by the microscope results in unusable sections. This function
takes strings such as "topRight" and "center" and returns pixel
boundaries based on those strings.
}

