% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow_labelDye.R
\name{flow_labelDye}
\alias{flow_labelDye}
\title{Identify and Label Florescent Dye}
\usage{
flow_labelDye(image, phase.labeled, labelingThreshold = 0.9)
}
\arguments{
\item{image}{an image matrix to search for dye}

\item{phase.labeled}{a labeled phase image from \link{flow_labelPhase}}

\item{labelingThreshold}{threshold value for detection [0-1]}
}
\value{
A \code{matrix} of integer labeled blobs.
}
\description{
Searches an image for dark cell colonies and incrementally labels each blob.
}

