% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow_loadImages.R
\name{flow_loadImages}
\alias{flow_loadImages}
\title{Load Image Data into a List}
\usage{
flow_loadImages(dir, ext = "tiff", start = 1, n = NA)
}
\arguments{
\item{dir}{path to the directory containing images}

\item{ext}{file extension}

\item{start}{index at which to start loading files}

\item{n}{number of images to load into memory (defaults to all)}
}
\value{
A \code{list} with two attributes, "images" and "filenames".
}
\description{
This function uses the \pkg{EBImage::readimage} function
to read in a series of images. Filenames for images are assumed to be
ordered (e.g. with numeric indices). The return list contains
two attributes:
\itemize{
\item{images -- a \code{list} of image matrices extracted from the \code{EBImage}
\code{image} object.}
\item{filenames -- a vector of filenames that correponds
to the images.}
}

The individual images are simple 2D matrices.
}
\examples{
\dontrun{
phase <- loadImages(dataDirPhase, n=30, ext="tif")
}
}

