% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/formatTime.R
\name{formatTime}
\alias{formatTime}
\title{Formats Elapsed Time}
\usage{
formatTime(proctime, digits = 2)
}
\arguments{
\item{proctime}{a proc_time object}

\item{digits}{the number of digits to round the output time to}
}
\value{
a formated time with "seconds" appended at the send.
}
\description{
Internal function used for debugging and printing output.
Cleans up the elapsed time of a proc_time object and prints it in
a format consistent with the rest of the output stream.
}

