% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateBlobTimeseries.R
\name{generateBlobTimeseries}
\alias{generateBlobTimeseries}
\title{Searches an image for dark cell colonies and incrementally labels each colony.}
\usage{
generateBlobTimeseries(frames, ignore = list(), minTimespan = 5,
  maxDistance = 20)
}
\arguments{
\item{frames}{a sequence of labeled frames. See  \link{labelColonies}.}

\item{ignore}{a list of y index pairs to ignore. Typically used to remove
centroids near dark lines.}

\item{minTimespan}{remove blobs from output which aren't found in at least
n sequential frames.}
}
\value{
A \code{list} with elements \code{timeseries}, a dataframe of blob IDs
and blob sizes at each timestep (in pixels) and \code{centroids}, a \code{list}
of dataframes with centroids, blob ID's and original integer labels for mapping
output column names back to the original images.
}
\description{
Searches an image for dark cell colonies and incrementally labels each colony.
}

