% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCentroids.R
\name{getCentroids}
\alias{getCentroids}
\title{Finds Blob Centroids}
\usage{
getCentroids(image)
}
\arguments{
\item{image}{a labeled image to search}
}
\value{
A \code{dataframe} with a row for each unique blob.
}
\description{
Finds the center, size, and bounds of every non zero cluster
of pixels. Requires the image was labeled using \code{EBImage::bwlabel}.
}

