% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TBCellGrowth.R
\name{getProfileSecs}
\alias{getProfileSecs}
\title{Get Profile Secs}
\usage{
getProfileSecs()
}
\value{
List of profiling stats.
}
\description{
Returns a named list containing the cumulative seconds associated
with a variety of tasks as determend by calls to profilePoint().
}
\seealso{
profilePoint
}
\keyword{environment}

