% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TBCellGrowth.R
\name{getRunOptions}
\alias{getRunOptions}
\title{Get Run Options}
\usage{
getRunOptions(option = NULL)
}
\arguments{
\item{option}{name of a specific option}
}
\value{
List of command line flags.
}
\description{
Return the value of a particular option or, if \code{param} is not specified,
a list of command line options such as \code{verbose} and \code{profile}.
}
\keyword{environment}

