% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cropImageByID.R
\docType{data}
\name{id}
\alias{id}
\title{Crops an Image Around a Given Blob ID}
\format{\preformatted{ chr "x1112y476z68"
}}
\usage{
id
}
\arguments{
\item{color}{a color string, i.e. "green" or "phase"}

\item{bg}{a background image}

\item{label}{a binary image label}
}
\value{
a color Image object
}
\description{
Given a blob ID string, crops an image around that id.
}
\keyword{datasets}

