% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/labelGroupsGreen.R
\name{labelGroupsGreen}
\alias{labelGroupsGreen}
\title{Identify and label Green Dye Groups}
\usage{
labelGroupsGreen(image, artifactMask)
}
\arguments{
\item{image}{an image matrix to search for green dye}

\item{artifactMask}{a mask of non biological features to ignore. See \link{createArtifactMask}.}
}
\value{
A \code{matrix} of integer labeled blobs.
}
\description{
Searches an image for dark cell colonies and incrementally labels each colony.
}

