% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadFrames.R
\name{loadFrames}
\alias{loadFrames}
\title{Load frames from a directory.}
\usage{
loadFrames(dir, ext, n = NA)
}
\arguments{
\item{dir}{A path to the directory containing image frames.}

\item{ext}{file extension, "tiff" by default.}

\item{n}{The number of frames to load into memory. Loads all by default.}
}
\value{
A \code{list} of \code{matrices} extracted from the \code{EBImage}
\code{image} object.
}
\description{
Load frames from a directory.
}

