% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadImages.R
\name{loadImages}
\alias{loadImages}
\title{Load Images into a List}
\usage{
loadImages(dir, ext = "tiff", n = NA)
}
\arguments{
\item{dir}{path to the directory containing images}

\item{ext}{file extension}

\item{n}{number of images to load into memory (defaults to all)}
}
\value{
A \code{list} of \code{matrices} extracted from the \code{EBImage}
\code{image} object.
}
\description{
Load Images into a List
}

