% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TBCellGrowth.R
\docType{data}
\name{nameList}
\alias{nameList}
\title{First Names}
\format{A character vector of length 4368.}
\description{
A vector of 4368 unique first names obtained from (...). These names
are used to identify individual cell colonies being tracked. First names are much
easier to remember than IDs generated from a hashing algorithm or numeric XY
coordinates.
}
\keyword{datasets}

