% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlayAnnotations.R
\name{overlayAnnotations}
\alias{overlayAnnotations}
\title{Add Annotations to an Image}
\usage{
overlayAnnotations(image, labeledImage, id, infoList)
}
\arguments{
\item{image}{an image matrix}

\item{labeledImage}{a matrix of blob labels}

\item{id}{the blob id the image is focused on}

\item{infoList}{list of information to be used for annotations}
}
\value{
a \code{matrix} image.
}
\description{
Adds a bar to the bottom right of an image.
The bar is labeled according to the scale that is passed in.
}

