% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlayBlobs.R
\name{overlayBlobs}
\alias{overlayBlobs}
\title{Overlays white blobs over an image.}
\usage{
overlayBlobs(image, blobLabels)
}
\arguments{
\item{image}{the image to modify.}

\item{blobLabels}{an equal sized image generated by \link{labelColonies}.}
}
\value{
an image of the same dimensions.
}
\description{
Masks an image with white blobs.
}

