% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlayColor.R
\name{overlayColor}
\alias{overlayColor}
\title{Adds a Colored Overlay to an Image}
\usage{
overlayColor(color, bg, label, mergeWith = NULL)
}
\arguments{
\item{color}{a color string, i.e. "green" or "phase"}

\item{bg}{a background image}

\item{label}{a binary image label}

\item{merge}{an equally sized color image to overlay the new color on. Uses background
image as default}
}
\value{
a color Image object
}
\description{
Adds a colored overlay to a background image. When passed "phase"
for color uses grey by default
}

