% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlayDyes.R
\name{overlayDyes}
\alias{overlayDyes}
\title{Overlay Dye Colors on an Image}
\usage{
overlayDyes(phase, green, red)
}
\arguments{
\item{phase}{a preprocessed phase image}

\item{green}{a labeled green dye image}

\item{red}{a labeled red dye image}
}
\value{
a color Image object
}
\description{
Overlays green and red labeled dye images on a phase image.
}

