% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlayGrid.R
\name{overlayGrid}
\alias{overlayGrid}
\title{Adds a grid to an image.}
\usage{
overlayGrid(image, spacing = 200)
}
\arguments{
\item{image}{the image to modify.}

\item{spacing}{how many pixels between each grid line.}
}
\value{
an image of equal dimensions.
}
\description{
Adds a white grid to an image with 1px lines. Grid labels are on the left and top.
}

