% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlayIDs.R
\name{overlayIDs}
\alias{overlayIDs}
\title{Adds colony ID text labels to a given image.}
\usage{
overlayIDs(image, centroids, labelbg = FALSE)
}
\arguments{
\item{image}{the image to modify.}

\item{centroids}{a \code{dataframe} corresponding to this image.
See \link{generateBobTimeseries}.}

\item{labelbg}{if TRUE adds a transparent background to make the labels
stand out more.}
}
\value{
an image of equal dimensions.
}
\description{
Uses the centroids data for a given image to label blob IDs directly
above blobs.
}

