% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlayNeighborIDs.R
\name{overlayNeighborsIDs}
\alias{overlayNeighborsIDs}
\title{Adds ID Labels to Neighboring BLobs}
\usage{
overlayNeighborsIDs(image, labels, id, centroids)
}
\arguments{
\item{image}{an image matrix}

\item{labels}{a matrix of blob labels}

\item{id}{the blob id the image is focused on}

\item{centroids}{a dataframe of blob centroids corresponding to this frame}
}
\value{
a \code{matrix} image.
}
\description{
Add text id labels to nearby blobs for cropped images.
}

