% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlayScaleBar.R
\name{overlayScaleBar}
\alias{overlayScaleBar}
\title{Add a Scale Bar to an Image}
\usage{
overlayScaleBar(image, distanceScale, thresholds = c(10, 50, 100))
}
\arguments{
\item{image}{an image matrix}

\item{distanceScale}{the image scale in micrometers}

\item{thresholds}{the length of the scale bar in sequential order. The last
number in the vector will be used unless the image is small enough for one
of the other values.}
}
\value{
a \code{matrix} image.
}
\description{
Adds a bar to the bottom right of an image.
The bar is labeled according to the scale that is passed in.
}

