% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlayVitalStats.R
\name{overlayVitalStats}
\alias{overlayVitalStats}
\title{Adds Basic Statistics to an Image}
\usage{
overlayVitalStats(image, id, time, size, distanceScale)
}
\arguments{
\item{image}{an image matrix}

\item{id}{the blob id the image is focused on}

\item{time}{the timestep at this frame}

\item{size}{the object size in um}

\item{distanceScale}{the distance conversion in um/pixel}
}
\value{
a \code{matrix} image.
}
\description{
Overylay a series of statistics to an image
from cropImageByID. TODO add more details when they are
avaiable
}

