% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotToOverlay.R
\name{plotToOverlay}
\alias{plotToOverlay}
\title{Convert Plot Function to Overlay Mask}
\usage{
plotToOverlay(plotf, dimx, dimy)
}
\arguments{
\item{plotf}{a function which plots lines, points, or text over the blank plot}

\item{dimx}{the image's x dimension}

\item{dimy}{the image's y dimension}
}
\value{
a \code{matrix} image.
}
\description{
This function runs plotting commands in \code{plotf} to add to an
otherwise blank plot of dimensions \code{dimx} X \code{dimy}. This image is
written to disk and then read in with \code{EBImage::readImage} to create a
a 2D mask that is 1 for each pixel that is part of the lines, points or text
}

