% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocessImages.R
\name{preprocessFrames}
\alias{preprocessFrames}
\title{Align, normalize brightness, and crop a list of image matrices.}
\usage{
preprocessFrames(frames, rotation = 0, sample = c(150, 450, 125),
  crop = c(150, 150, 150, 150))
}
\arguments{
\item{frames}{a list of matrices.}

\item{rotation}{how many degrees to rotate each image.}

\item{sample}{which region of the frames to use for alignment.}

\item{crop}{amount to crop from each side in format \code{c(bottom, left, top,
right)}.}
}
\value{
A \code{list} of \code{matrices} ready for feature extraction.
}
\description{
Align, normalize brightness, and crop a list of image matrices.
}

