% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocessImages.R
\name{preprocessImages}
\alias{preprocessImages}
\title{Align, Normalize Brightness, and Crop a List of Image Matrices}
\usage{
preprocessImages(images, rotation = 0, sample = c(150, 450, 125),
  crop = c(150, 150, 150, 150))
}
\arguments{
\item{images}{a list of image matrices}

\item{rotation}{how many degrees to rotate each image}

\item{sample}{which region of the images to use for alignment}

\item{crop}{amount to crop from each side in format \code{c(bottom, left, top, right)}.}
}
\value{
A \code{list} of \code{matrices} ready for feature extraction.
}
\description{
Align, Normalize Brightness, and Crop a List of Image Matrices
}

