% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TBCellGrowth.R
\name{profileEnd}
\alias{profileEnd}
\title{Print Total Elapsed Time}
\usage{
profileEnd(message = NULL)
}
\arguments{
\item{message}{character message}
}
\value{
Elapsed time in seconds.
}
\description{
Returns the elapsed time since profileStart().
If \code{message} is not \code{NULL}, a message with this information is printed out.
}
\seealso{
profileStart

profilePoint
}
\keyword{environment}

