% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TBCellGrowth.R
\name{profilePoint}
\alias{profilePoint}
\title{Print Partial Elapsed Time}
\usage{
profilePoint(name = "unknown", message = NULL)
}
\arguments{
\item{name}{named counter which will be incremented with secs since last timepoint}

\item{message}{character message}
}
\value{
Elapsed time in seconds (invisbly).
}
\description{
Returns the elapsed time since the
last time profilePoint() was called or since profileStart().
If \code{message} is not \code{NULL}, an message with this information is printed out.
}
\seealso{
profileStart

profileEnd
}
\keyword{environment}

