% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TBCellGrowth.R
\name{profileStart}
\alias{profileStart}
\title{Start Profiling Timers}
\usage{
profileStart()
}
\value{
None
}
\description{
Sets internal *start* and *timepoint* times that
can be used in scripts to calculate elapsed 'clock' time.
}
\seealso{
profilePoint

profileEnd
}
\keyword{environment}

