% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/saveImageList.R
\name{saveImageList}
\alias{saveImageList}
\title{Save List of Images as Files}
\usage{
saveImageList(imageList, outputDir, chamber, tag)
}
\arguments{
\item{imageList}{list of images orgainzed as [[channel]][[timestep]]}

\item{outputDir}{output directory}

\item{chamber}{name of the chamber whose images are loaded e.g. "xy2"}

\item{tag}{identifier appended to file names}
}
\value{
none
}
\description{
Images in \code{imageList} are written to disk with the
following filename: \code{chamber_channel_timestep_tag.jpg}.
}

