% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TBCellGrowth.R
\name{setRunOptions}
\alias{setRunOptions}
\title{Set Run Options}
\usage{
setRunOptions(opt)
}
\arguments{
\item{opt}{list of options}
}
\value{
None
}
\description{
Store command line options such as \code{verbose} and \code{profile}
so that they are available internally.
}
\keyword{environment}

