% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/solid_alignImages.R
\name{solid_alignImages}
\alias{solid_alignImages}
\title{Align A Series of Flow Images}
\usage{
solid_alignImages(images, numTargets = 12, targetWidth = 50,
  searchSpace = 50)
}
\arguments{
\item{images}{a list of lists of images with channel names as keys.}

\item{numTargets}{how many target features to compare}

\item{targetWidth}{the radius of alignment targets}

\item{searchSpace}{how far to search for best alignment. A smaller
value here yeilds faster searches.}
}
\value{
a \code{list} of two lists, \code{phase} and \code{dyes},
which will be the same lengths as the input.
}
\description{
Given a list of phase images and a list of zero or
more lists of dye images, aligns these images to the first image
of phase (assumed to be a background). Will also crop and rotate
the images if those are specified.
}

