% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solid_equalizeDye.R
\name{solid_equalizeDye}
\alias{solid_equalizeDye}
\title{Normalize Dye Image Values}
\usage{
solid_equalizeDye(image)
}
\arguments{
\item{image}{the image matrix to modify.}
}
\value{
an image of the same dimensions.
}
\description{
Normalize a given dye image. TODO More info here
}

