% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/solid_equalizeDyeImages.R
\name{solid_equalizeDyeImages}
\alias{solid_equalizeDyeImages}
\title{Normalize Dye Image Values}
\usage{
solid_equalizeDyeImages(image)
}
\arguments{
\item{image}{the image matrix to modify.}

\item{artifactMask}{a mask of non biological features to ignore. See \link{flow_createArtifactMask}.}
}
\value{
an image of the same dimensions.
}
\description{
Normalize a given dye image. TODO More info here
}

