% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solid_equalizePhase.R
\name{solid_equalizePhase}
\alias{solid_equalizePhase}
\title{Normalize Image Values}
\usage{
solid_equalizePhase(image)
}
\arguments{
\item{image}{the image matrix to modify.}
}
\value{
an image of the same dimensions.
}
\description{
Normalize a given solid image. This function attempts to
replicate the auto levels function in Photoshop. It shifts the peak of
the histogram to 0 and scales the values so the histogram is between
0 and ~0.5
}

