% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/solid_labelDye.R
\name{solid_labelDye}
\alias{solid_labelDye}
\title{Identify and Label Florescent Dye}
\usage{
solid_labelDye(image, phase.labeled)
}
\arguments{
\item{image}{an image matrix to search for dye}

\item{phase.labeled}{a list of labeled phase images from
\link{flow_labelPhase}}
}
\value{
A \code{matrix} of integer labeled blobs.
}
\description{
Searches an image for dark cell colonies and incrementally labels each blob.
}

