% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/solid_loadImages.R
\name{solid_loadImages}
\alias{solid_loadImages}
\title{Load Image Data into a List}
\usage{
solid_loadImages(dataDir, xy, channels = c("c1"), channelNames = c("phase"),
  ext = "tiff", start = 1, n = NA)
}
\arguments{
\item{dataDir}{path to the directory containing images}

\item{xy}{the xy number to load e.g. "xy2"}

\item{channels}{which channels to load for the given xy}

\item{channelNames}{a vector of names corresponding to the
given channels}

\item{ext}{file extension}

\item{start}{index at which to start loading files}

\item{n}{number of images to load into memory (defaults to all)}
}
\description{
This function uses the \pkg{EBImage::readimage} function
to read in a series of images. Filenames for images are assumed to be
ordered (e.g. with numeric indices). The return is a list of image
matrices from the \code{.Data} slot of the EBImage Image object.
}

