\name{TBCellGrowth-package}
\alias{TBCellGrowth-package}
\alias{TBCellGrowth}
\docType{package}
\title{
\packageTitle{TBCellGrowth}
}
\description{
\packageDescription{TBCellGrowth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TBCellGrowth}
\packageIndices{TBCellGrowth}
}
\author{
\packageAuthor{TBCellGrowth}

Maintainer: \packageMaintainer{TBCellGrowth}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
