% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TBCellGrowth.R
\docType{package}
\name{TBCellGrowth}
\alias{TBCellGrowth}
\alias{TBCellGrowth-package}
\title{Cell Microscopy Feature Extraction and Growth Tracking}
\description{
A suite of functions for processing cell microscopy data,
specifically developed for tuberculosis. Features include GLCM based
feature extraction and blob labeling as well as tracking of colony growth
over many frames. Cell growth output is available as CSV as well as a
variety of image overlays and animations.
}

