% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeExcel.R
\name{writeExcel}
\alias{writeExcel}
\title{Create Excel File from Dataframe}
\usage{
writeExcel(df, outputDir, channel, filenames, chamber)
}
\arguments{
\item{df}{timeseries dataframe created by \link{generateBlobTimeseries}}

\item{outputDir}{output directory}

\item{channel}{name of the channel e.g. "phase" or "green"}

\item{filenames}{a vector of names equal to the number of rows of the dataframe.
Typically timesteps.}

\item{chamber}{name of the chamber whose images are loaded e.g. "xy2"}
}
\value{
Absolute path of the "no links" version of the .csv file.
}
\description{
A specially formatted .csv file is generated from the contents
of the \code{df} dataframe. Individual cells will be linked to images associated
with each TB colony (row) at each timestep (col). Full columns will be linked
to animaged gifs while full rows will be linked to full images for each timestep.

The images are generated in buildDirectoryStructure() which also calls this function.
}

