% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeImages.R
\name{writeImages}
\alias{writeImages}
\title{Creates Images for Excel Hyperlinks}
\usage{
writeImages(images, outputDir, id, channel, filenames)
}
\arguments{
\item{images}{list of images}

\item{outputDir}{the directory to build the file structure}

\item{id}{cell colony name (column headers in the Excel file)}

\item{channel}{channel name e.g. "phase","green","red"}

\item{filenames}{a vector of names equal to the length of the
image lists. Usually timesteps.}
}
\value{
none
}
\description{
Creates a subdirectory based on \code{outputDir},
\code{id} and \code{channel} and then populates that directory
with the .jpg versions of the \code{images}, each labeled
according to the associated \code{filenames}.

Hyperlinks for these images are generated in writeExcel().
}

