% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeIndividualImages.R
\name{writeIndividualImages}
\alias{writeIndividualImages}
\title{Write Individual Images}
\usage{
writeIndividualImages(timeseriesList, phase, labeledImageList, dyeOverlap,
  filenames, outputDir = "output", distanceScale = NULL, chamber = "")
}
\arguments{
\item{timeseriesList}{a List with timeseries and centroids from
\link{generateBlobTimeseries}}

\item{phase}{a list of phase contrast image matrices}

\item{labeledImageList}{a list of lists of labeled images for each channel
\link{flow_labelPhase}}

\item{dyeOverlap}{a \code{dataframe} of dye overlap values from
\link{findDyeOverlap}}

\item{filenames}{a vector of names equal to the length of the
image lists. Usually timesteps.}

\item{outputDir}{the directory to build the file structure}

\item{distanceScale}{the distance conversion in um/pixel}

\item{chamber}{the current chamber id}
}
\value{
none
}
\description{
Creates individual directories and populates each
with colony thumbnail images with outlines and other labeling.
}

