% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis_doublingTime.R
\name{analysis_doublingTime}
\alias{analysis_doublingTime}
\title{Calculate Doubling Times}
\usage{
analysis_doublingTime(timeseries, minExpFitHour = 0, maxExpFitHour = 1e+09)
}
\arguments{
\item{timeseries}{timeseries dataframe}

\item{minExpFitHour}{hour of first datapoint to include in doubling time exponential fit}

\item{maxExpFitHour}{hour of last datapoint to include in doubling time exponential fit}
}
\value{
Vector of doubling times
}
\description{
Given a "timeseries" dataframe with hours in the
first column and colony sizes in all other columns, return a
vector of doubling times in hours.
}
\details{
The algorithm for calculating doubling times is very straightforward:
\preformatted{
  1) hours = first column of incoming 'timeseries' dataframe
  2) colonies = all other columns of incoming 'timeseries' dataframe
  3) subset both 'hours' and 'colonies' to include only the desired rows (hours)
  4) for each column in 'colonies':
       slope = linear least squares regression of ln(colony_size) vs. hours
       doubling time = log(2) / slope
}
}
\references{
http://www.pangloss.com/wiki/R_GrowthCurve
}

