% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis_doublingTime.R
\name{analysis_doublingTime}
\alias{analysis_doublingTime}
\title{Calculate Doubling Times}
\usage{
analysis_doublingTime(timeseries, rowStart = 1, rowStop = nrow(timeseries))
}
\arguments{
\item{timeseries}{timeseries dataframe}

\item{rowStart}{index of starting hour}

\item{rowStop}{index of ending hour}
}
\value{
Vector of doubling times
}
\description{
Given a "timeseries" dataframe with hours in the
first column and colony sizes in all other columns, return a
vector of doubling times in hours.
}

