% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis_twoPlot.R
\name{analysis_twoPlot}
\alias{analysis_twoPlot}
\title{Two Plots of Growth Curves}
\usage{
analysis_twoPlot(timeseries, rowStart = 1, rowStop = nrow(timeseries),
  title = "Title Goes Here", filename = NULL, pngSize = 800)
}
\arguments{
\item{timeseries}{timeseries dataframe}

\item{rowStart}{index of starting hour}

\item{rowStop}{index of ending hour}

\item{title}{overall title}

\item{filename}{path of the file to which the plot is saved (NULL will plot to screen)}

\item{pngSize}{image width/height in pixels}
}
\value{
none
}
\description{
Given a "timeseries" dataframe with hours in the
first column and colony sizes in all other columns, create growth curve plots for
all columns.
}

