% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/analysis_winnowColonies.R
\name{analysis_winnowColonies}
\alias{analysis_winnowColonies}
\title{Remove Likely Artifacts}
\usage{
analysis_winnowColonies(timeseries, maxDoublingTime = NULL,
  removeOutliers = TRUE, minStartHour = NULL)
}
\arguments{
\item{timeseries}{timeseries dataframe}

\item{maxDoublingTime}{colonies with doubling times longer than this are removed}

\item{removeOutliers}{flag indicating whether to remove colonies with doubling time outliers}

\item{minStartHour}{colonies not identified by this hour are removed}
}
\value{
List of two timeseries dataframes: "retained" and "removed".
}
\description{
After image analysis, the TBCellGrowth package produces.csv files
which can be read in as a "timeseries" dataframe
with hours in the first column and colony sizes in all other columns.
Depending on run configuration, some portion of the colonies will be artifacts
-- bubbles, blotches, specks, etc.

This function helps to winnow down liktely artifacts and returns a dataframe
that likely contains a higher percentage of actual colonies.

The main way this function removes artifacts is by removing "colonies" that
do not grow. Doubling times for rapidly growing colonies will be low numbers
while doubling times for no or little growht will be very large numbers.
Negative doubling times are found where identified "colonies" shrink over time.

This function identifies any colony with a negative doubling time as an artifact.
After removal of negative doubling times, the boxlplot() fucntion is used to
identifiy outliers. These are subsequently removed.

An additional option is to remove all "colonies" that are not identified until
after a particular hour. (Sometimes, changes in the image quality cause new
"colonies" to be identified after a particular timestep.)

This funciton returns a list with two dataframes, one containing the retained
colonies and one containing the removed colonies.
}

