% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/buildDirectoryStructure.R
\name{buildDirectoryStructure}
\alias{buildDirectoryStructure}
\title{Builds A Directory of Image With Excel UI}
\usage{
buildDirectoryStructure(output, phase, labeled, dyeOverlap, filenames,
  outputDir = "output", distanceScale = NULL)
}
\arguments{
\item{output}{a List with timeseries and centroids from
\link{generateBlobTimeseries}}

\item{phase}{a list of phase contrast image matrices}

\item{labeled}{a list of labeled images from each channel
\link{flow_labelPhase}}

\item{dyeOverlap}{a \code{dataframe} of dye overlap values from
\link{findDyeOverlap}}

\item{filenames}{a vector of names equal to the length of the
image lists. Usually timesteps.}

\item{outputDir}{the directory to build the file structure}

\item{distanceScale}{the distance conversion in um/pixel}
}
\value{
none
}
\description{
write a lot more here TODO
}

