% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createGif.R
\name{createGif}
\alias{createGif}
\title{Creates a gif animation.}
\usage{
createGif(frames, filename, delay = 15, rescale = 100)
}
\arguments{
\item{frames}{a sequential \code{list} of images.}

\item{filename}{name of output gif.}

\item{delay}{delay between each frame in milliseconds.}

\item{rescale}{dimensions of gif frames as a percent of original image size.}
}
\value{
none
}
\description{
Creates an animated .gif file from a list of images using ImageMagick.
}

