% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cropImageByID.R
\name{cropImageByID}
\alias{cropImageByID}
\title{Crops an Image Around a Given Blob ID}
\usage{
cropImageByID(id, centroids, phase, labeled)
}
\arguments{
\item{id}{the blob id to crop image around}

\item{centroids}{a centroids dataframe as generated by
\link{generateBlobTimeseries}}

\item{phase}{a list of phase image matrices, one per timestep}

\item{labeled}{a binary image label}
}
\value{
a list with elements bg, the cropped background, and label, the cropped
label
}
\description{
Given a blob ID string, look through all centroids for that
image and calculate a reasonable rectangle and then crop images around that ID.
}

