% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cropImageByID.R
\name{cropImageByID}
\alias{cropImageByID}
\title{Crops an Image Around a Given Blob ID}
\usage{
cropImageByID(id, output, bg, label)
}
\arguments{
\item{id}{the blob id to crop image around}

\item{output}{an output object with timeseries and centroids from
\link{generateBlobTimeseries}}

\item{bg}{a phase image matrix}

\item{label}{a binary image label}
}
\value{
a list with elements bg, the cropped background, and label, the cropped
label
}
\description{
Given a blob ID string, crops an image around that id.
}

