% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findDarkLines.R
\name{findDarkLines}
\alias{findDarkLines}
\title{Find horizontal dark lines, typically the space between plate levels.}
\usage{
findDarkLines(m)
}
\arguments{
\item{m}{an image matrix to search for dark lines.}
}
\value{
A \code{dataframe} of start and end indices for dark lines.
}
\description{
Find horizontal dark lines, typically the space between plate levels.
}

