% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findDyeOverlap.R
\name{findDyeOverlap}
\alias{findDyeOverlap}
\title{Find Overlap Between Dye and Phase Blobs}
\usage{
findDyeOverlap(imageList, timeseriesList)
}
\arguments{
\item{imageList}{a list of labeled dye images}

\item{timeseriesList}{a list with timeseries and centroids from \link{generateBlobTimeseries}}
}
\value{
A \code{matrix} of integer labeled blobs.
}
\description{
For every labeled blob in every dye image, count the number of pixels
where the blob overlaps with a blob generated from the phase images.
}

