% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow_alignImages.R
\name{flow_alignImages}
\alias{flow_alignImages}
\title{Align A Series of Flow Images}
\usage{
flow_alignImages(images, alignmentTargets, targetWidth = 25,
  searchSpace = 25)
}
\arguments{
\item{images}{a list of lists of images with channel names as keys.}

\item{alignmentTargets}{a list of x y coordinates to use for alignment
comparisons}

\item{targetWidth}{the radius of alignment targets}

\item{searchSpace}{how far to search for best alignment. A smaller
value here yeilds faster searches.}
}
\value{
a \code{list} of two lists, \code{phase} and \code{dyes},
which will be the same lengths as the input.
}
\description{
Given a list of phase images and a list of zero or
more lists of dye images, aligns these images to the first image
of phase (assumed to be a background). Will also crop and rotate
the images if those are specified.
}

