% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow_equalizeDye.R
\name{flow_equalizeDye}
\alias{flow_equalizeDye}
\title{Normalize Dye Image Values}
\usage{
flow_equalizeDye(image, artifactMask)
}
\arguments{
\item{image}{the image matrix to modify.}

\item{artifactMask}{a mask of non biological features to ignore. See \link{flow_createArtifactMask}.}
}
\value{
an image of the same dimensions.
}
\description{
Normalize a given dye image. TODO More info here
}

