% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_equalizeDyeImages.R
\name{flow_equalizeDyeImages}
\alias{flow_equalizeDyeImages}
\title{Normalize Dye Image Values}
\usage{
flow_equalizeDyeImages(im, artifactMask)
}
\arguments{
\item{image}{the image matrix to modify.}

\item{medianNew}{a new median value to set scale the image values against}
}
\value{
an image of the same dimensions.
}
\description{
Normalize a given dye image. TODO More info here
}

