% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_labelDye.R
\name{flow_labelDye}
\alias{flow_labelDye}
\title{Identify and Label Florescent Dye}
\usage{
flow_labelDye(image, phase.labeled, artifactMask)
}
\arguments{
\item{image}{an image matrix to search for dye}

\item{artifactMask}{a mask of non biological features to ignore. See \link{createArtifactMask}.}
}
\value{
A \code{matrix} of integer labeled blobs.
}
\description{
Searches an image for dark cell colonies and incrementally labels each blob.
}

