% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_rotateImages.R
\name{flow_rotateImages}
\alias{flow_rotateImages}
\title{Rotate an Image}
\usage{
flow_rotateImages(imageList, searchSpace = seq(-2, 2, by = 0.2))
}
\arguments{
\item{imageList}{list of images orgainzed as [[channel]][[timestep]]}

\item{searchSpace}{a vector of rotation degrees to check.}
}
\value{
A imageList with images of equal or smaller dimensions to the input.
}
\description{
Uses the EBImage rotate function to rotate a list of images.
The 'flow' images typically have horizontal lines of bright or dark pixels.
As the background images is rotated through a series of angles, the row
averages are calculated. The standard deviation of these row averages is
the calculated and the rotation that maximizes this standard deviation
is chosen as the optimal rotation. This will occur when while and dark
lines are horizontally aligned.

After rotation, images will be cropped to remove the excess black space
introduced by the rotation.
}

