% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow_rotateImages.R
\name{flow_rotateImages}
\alias{flow_rotateImages}
\title{Rotate an Image}
\usage{
flow_rotateImages(phase, dyes, searchSpace = seq(-2, 2, by = 0.2))
}
\arguments{
\item{phase}{a list of phase images}

\item{dyes}{a list of lists dye images of equal length to phase}

\item{searchSpace}{a vector of rotation degrees to check.}
}
\value{
an image of equal or smaller dimensions to the input.
}
\description{
Uses the EBImage rotate function to rotate a series
of dye and phase images. An algorithm determines which degree value
in the searchSpace will be applied. The image will also be cropped
to remove the excess black space introduced by the rotate
}

