% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateBlobTimeseries.R
\name{generateBlobTimeseries}
\alias{generateBlobTimeseries}
\title{Searches an image for dark cell colonies and incrementally labels each colony.}
\usage{
generateBlobTimeseries(images, minTimespan = 8, maxDistance = 50)
}
\arguments{
\item{images}{a list of labeled images See \link{flow_labelPhase}.}

\item{minTimespan}{remove blobs from output which aren't found in at least
n sequential images}

\item{maxDistance}{the cutoff for distance between two blobs}
}
\value{
A \code{list} with elements \code{timeseries}, a dataframe of blob IDs
and blob sizes at each timestep (in pixels) and \code{centroids}, a \code{list}
of dataimages with centroids, blob ID's and original integer labels for mapping
output column names back to the original images
}
\description{
Searches an image for dark cell colonies and incrementally labels each colony.
}

