% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loadImages.R
\name{loadImages}
\alias{loadImages}
\title{Load Image Data into a List}
\usage{
loadImages(dataDir, chamber, channels = c("c1"), channelNames = c("phase"),
  ext = "tiff", startFrame = 1, n = "all")
}
\arguments{
\item{dataDir}{path to the directory containing images}

\item{chamber}{name of the chamber whose images are loaded e.g. "xy2"}

\item{channels}{which channels to load for the given chamber}

\item{channelNames}{a vector of names to be associated with the the given channels}

\item{ext}{file extension}

\item{startFrame}{index at which to start loading files}

\item{n}{number of images to load into memory}
}
\value{
A list of image matrices.
}
\description{
This function uses the \pkg{EBImage::readimage} function
to read in a series of images. Filenames for images are assumed to be
ordered (e.g. with numeric indices). The return is a list of two multi-level lists.

The \code{imageList} is a list of image matrices obtained by reading in files with \code{EBImage::readImage()}
and then extractng the \code{.Data} slot of the EBImage Image object.
Images that could not be read in have an NA stored in their slot.

The organizational structure is:

\code{imageList[[channel]][[timestep]]}
}

