% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlayOutlines.R
\name{overlayOutlines}
\alias{overlayOutlines}
\title{Overlays Colored Outlines Over Blobs}
\usage{
overlayOutlines(image, mask, color = "red", thick = TRUE)
}
\arguments{
\item{image}{the image matrix to modify.}

\item{mask}{an equal sized image generated by a label function,
e.g. \link{flow_labelPhase}.}

\item{color}{the outline color}

\item{thick}{if TRUE then outlines will be thickened}
}
\value{
an image of the same dimensions.
}
\description{
Given an image matrix and a mask of equal dimension, draws
outlines around the mask blobs on the image. Assumes image and mask are
of equal dimensions.
}

