% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overlayScaleBar.R
\name{overlayScaleBar}
\alias{overlayScaleBar}
\title{Add a Scale Bar to an Image}
\usage{
overlayScaleBar(image, distanceScale, barLength = 100)
}
\arguments{
\item{image}{an image matrix}

\item{distanceScale}{the image scale in micrometers}

\item{barLength}{how many pixels wide the scale bar should be}
}
\value{
a \code{matrix} image.
}
\description{
Adds a bar to the bottom right of an image.
The bar is labeled according to the scale that is passed in.
}

