% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotToOverlay.R
\name{plotToOverlay}
\alias{plotToOverlay}
\title{Saves a plot as an image and loads it as an image.}
\usage{
plotToOverlay(plotf, dimx, dimy)
}
\arguments{
\item{plotf}{a function which plots lines, points, or text over the blank plot.}

\item{dimx}{the image's x dimension.}

\item{dimy}{the image's y dimension.}
}
\value{
a \code{matrix} image.
}
\description{
Saves a plot as a png and opens it using EBImage.
}

