% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/removeBlobs.R
\name{removeBlobs}
\alias{removeBlobs}
\title{Remove Blobs By Size}
\usage{
removeBlobs(image, size)
}
\arguments{
\item{image}{an image matrix}

\item{size}{threshold for removal}
}
\value{
An image \code{matrix} of size equal to the input image
}
\description{
Finds the size of each blob group in an image and removes
blobs under the given threshold of pixels. A blob is defined as a group of
pixels connected horizontally or vertically.
}

